Code to run svscan on startup on mac and linux
specifically we install the startup scripts (rendered from the erbs)

Then, on a mac, we make sure that the appropriate paths are extant for
launchctl. If not, we examine the current paths, and install the new
paths, keeping the old paths sane.

we also write our new paths to /etc/launchd.conf so that they persist
after reboot.